Replace d3 with mapbox
Replace the svg map, d3, and related dependencies with the mapbox api.
The Mapbox map is rendered in a canvas outside React. We use the mapbox
api to position invisible markers at latlngs for each accident; we then
lookup the top and left offsets of each marker to create a visible
Marker component in React at that location.